[FLINK-24325][connectors/elasticsearch] Integrate common unified Elasticsearch sink with Table API for versions 6 and 7
