+++ flink-connectors/flink-connector-elasticsearch-base/src/main/java/org/apache/flink/streaming/connectors/elasticsearch/table/ElasticsearchValidationUtils.java
@@ -19,11 +19,9 @@
 package org.apache.flink.streaming.connectors.elasticsearch.table;
 
 import org.apache.flink.annotation.Internal;
-import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.ValidationException;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.DistinctType;
-import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.table.types.logical.LogicalTypeFamily;
 import org.apache.flink.table.types.logical.LogicalTypeRoot;
 
@@ -35,20 +33,9 @@ import java.util.stream.Collectors;
 /** Utility methods for validating Elasticsearch properties. */
 @Internal
 class ElasticsearchValidationUtils {
-
-    private static final Set<LogicalTypeRoot> ILLEGAL_PRIMARY_KEY_TYPES = new LinkedHashSet<>();
     private static final Set<LogicalTypeRoot> ALLOWED_PRIMARY_KEY_TYPES = new LinkedHashSet<>();
 
     static {
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.ARRAY);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.MAP);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.MULTISET);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.STRUCTURED_TYPE);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.ROW);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.RAW);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.BINARY);
-        ILLEGAL_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.VARBINARY);
-
         ALLOWED_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.CHAR);
         ALLOWED_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.VARCHAR);
         ALLOWED_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.BOOLEAN);
@@ -66,49 +53,6 @@ class ElasticsearchValidationUtils {
         ALLOWED_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.TIMESTAMP_WITH_LOCAL_TIME_ZONE);
         ALLOWED_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.INTERVAL_YEAR_MONTH);
         ALLOWED_PRIMARY_KEY_TYPES.add(LogicalTypeRoot.INTERVAL_DAY_TIME);
-    }
-
-    /**
-     * Checks that the table does not have primary key defined on illegal types. In Elasticsearch
-     * the primary key is used to calculate the Elasticsearch document id, which is a string of up
-     * to 512 bytes. It cannot have whitespaces. As of now it is calculated by concatenating the
-     * fields. Certain types do not have a good string representation to be used in this scenario.
-     * The illegal types are mostly {@link LogicalTypeFamily#COLLECTION} types and {@link
-     * LogicalTypeRoot#RAW} type.
-     */
-    @Deprecated
-    public static void validatePrimaryKey(TableSchema schema) {
-        schema.getPrimaryKey()
-                .ifPresent(
-                        key -> {
-                            List<LogicalTypeRoot> illegalTypes =
-                                    key.getColumns().stream()
-                                            .map(
-                                                    fieldName -> {
-                                                        LogicalType logicalType =
-                                                                schema.getFieldDataType(fieldName)
-                                                                        .get()
-                                                                        .getLogicalType();
-                                                        if (logicalType.is(
-                                                                LogicalTypeRoot.DISTINCT_TYPE)) {
-                                                            return ((DistinctType) logicalType)
-                                                                    .getSourceType()
-                                                                    .getTypeRoot();
-                                                        } else {
-                                                            return logicalType.getTypeRoot();
-                                                        }
-                                                    })
-                                            .filter(ILLEGAL_PRIMARY_KEY_TYPES::contains)
-                                            .collect(Collectors.toList());
-
-                            if (!illegalTypes.isEmpty()) {
-                                throw new ValidationException(
-                                        String.format(
-                                                "The table has a primary key on columns of illegal types: %s.\n"
-                                                        + " Elasticsearch sink does not support primary keys on columns of types: %s.",
-                                                illegalTypes, ILLEGAL_PRIMARY_KEY_TYPES));
-                            }
-                        });
     }
 
     /**
