+++ flink-connectors/flink-connector-elasticsearch6/src/main/java/org/apache/flink/streaming/connectors/elasticsearch/table/Elasticsearch6DynamicSinkFactory.java
@@ -19,152 +19,54 @@
 package org.apache.flink.streaming.connectors.elasticsearch.table;
 
 import org.apache.flink.annotation.Internal;
-import org.apache.flink.api.common.serialization.SerializationSchema;
 import org.apache.flink.configuration.ConfigOption;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.table.api.TableSchema;
-import org.apache.flink.table.api.ValidationException;
-import org.apache.flink.table.connector.format.EncodingFormat;
-import org.apache.flink.table.connector.sink.DynamicTableSink;
-import org.apache.flink.table.data.RowData;
+import org.apache.flink.connector.elasticsearch.sink.Elasticsearch6SinkBuilder;
 import org.apache.flink.table.factories.DynamicTableSinkFactory;
-import org.apache.flink.table.factories.FactoryUtil;
-import org.apache.flink.table.factories.SerializationFormatFactory;
-import org.apache.flink.table.utils.TableSchemaUtils;
 import org.apache.flink.util.StringUtils;
 
+import javax.annotation.Nullable;
+
 import java.util.Set;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.BULK_FLASH_MAX_SIZE_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.BULK_FLUSH_BACKOFF_DELAY_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.BULK_FLUSH_BACKOFF_MAX_RETRIES_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.BULK_FLUSH_BACKOFF_TYPE_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.BULK_FLUSH_INTERVAL_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.BULK_FLUSH_MAX_ACTIONS_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.CONNECTION_MAX_RETRY_TIMEOUT_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.CONNECTION_PATH_PREFIX;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.DOCUMENT_TYPE_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.FAILURE_HANDLER_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.FLUSH_ON_CHECKPOINT_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.FORMAT_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.HOSTS_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.INDEX_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.KEY_DELIMITER_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.PASSWORD_OPTION;
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.USERNAME_OPTION;
+import static org.apache.flink.streaming.connectors.elasticsearch.table.Elasticsearch6ConnectorOptions.DOCUMENT_TYPE_OPTION;
 
-/** A {@link DynamicTableSinkFactory} for discovering {@link Elasticsearch6DynamicSink}. */
+/** A {@link DynamicTableSinkFactory} for discovering {@link ElasticsearchDynamicSink}. */
 @Internal
-public class Elasticsearch6DynamicSinkFactory implements DynamicTableSinkFactory {
-    private static final Set<ConfigOption<?>> requiredOptions =
-            Stream.of(HOSTS_OPTION, INDEX_OPTION, DOCUMENT_TYPE_OPTION).collect(Collectors.toSet());
-    private static final Set<ConfigOption<?>> optionalOptions =
-            Stream.of(
-                            KEY_DELIMITER_OPTION,
-                            FAILURE_HANDLER_OPTION,
-                            FLUSH_ON_CHECKPOINT_OPTION,
-                            BULK_FLASH_MAX_SIZE_OPTION,
-                            BULK_FLUSH_MAX_ACTIONS_OPTION,
-                            BULK_FLUSH_INTERVAL_OPTION,
-                            BULK_FLUSH_BACKOFF_TYPE_OPTION,
-                            BULK_FLUSH_BACKOFF_MAX_RETRIES_OPTION,
-                            BULK_FLUSH_BACKOFF_DELAY_OPTION,
-                            CONNECTION_MAX_RETRY_TIMEOUT_OPTION,
-                            CONNECTION_PATH_PREFIX,
-                            FORMAT_OPTION,
-                            PASSWORD_OPTION,
-                            USERNAME_OPTION)
-                    .collect(Collectors.toSet());
+public class Elasticsearch6DynamicSinkFactory extends ElasticsearchDynamicSinkFactoryBase {
+    private static final String FACTORY_IDENTIFIER = "elasticsearch-6";
+
+    public Elasticsearch6DynamicSinkFactory() {
+        super(FACTORY_IDENTIFIER, Elasticsearch6SinkBuilder::new);
+    }
 
     @Override
-    public DynamicTableSink createDynamicTableSink(Context context) {
-        TableSchema tableSchema = context.getCatalogTable().getSchema();
-        ElasticsearchValidationUtils.validatePrimaryKey(tableSchema);
-        final FactoryUtil.TableFactoryHelper helper =
-                FactoryUtil.createTableFactoryHelper(this, context);
+    ElasticsearchConfiguration getConfiguration(Context context) {
+        return new Elasticsearch6Configuration(
+                Configuration.fromMap(context.getCatalogTable().getOptions()));
+    }
 
-        final EncodingFormat<SerializationSchema<RowData>> format =
-                helper.discoverEncodingFormat(SerializationFormatFactory.class, FORMAT_OPTION);
-
-        helper.validate();
-        Configuration configuration = new Configuration();
-        context.getCatalogTable().getOptions().forEach(configuration::setString);
+    @Nullable
+    @Override
+    String getDocumentType(Context context) {
         Elasticsearch6Configuration config =
-                new Elasticsearch6Configuration(configuration, context.getClassLoader());
-
-        validate(config, configuration);
-
-        return new Elasticsearch6DynamicSink(
-                format, config, TableSchemaUtils.getPhysicalSchema(tableSchema));
-    }
-
-    private void validate(Elasticsearch6Configuration config, Configuration originalConfiguration) {
-        config.getFailureHandler(); // checks if we can instantiate the custom failure handler
-        config.getHosts(); // validate hosts
-        validate(
-                config.getIndex().length() >= 1,
-                () -> String.format("'%s' must not be empty", INDEX_OPTION.key()));
-        int maxActions = config.getBulkFlushMaxActions();
-        validate(
-                maxActions == -1 || maxActions >= 1,
-                () ->
-                        String.format(
-                                "'%s' must be at least 1. Got: %s",
-                                BULK_FLUSH_MAX_ACTIONS_OPTION.key(), maxActions));
-        long maxSize = config.getBulkFlushMaxByteSize();
-        long mb1 = 1024 * 1024;
-        validate(
-                maxSize == -1 || (maxSize >= mb1 && maxSize % mb1 == 0),
-                () ->
-                        String.format(
-                                "'%s' must be in MB granularity. Got: %s",
-                                BULK_FLASH_MAX_SIZE_OPTION.key(),
-                                originalConfiguration
-                                        .get(BULK_FLASH_MAX_SIZE_OPTION)
-                                        .toHumanReadableString()));
-        validate(
-                config.getBulkFlushBackoffRetries().map(retries -> retries >= 1).orElse(true),
-                () ->
-                        String.format(
-                                "'%s' must be at least 1. Got: %s",
-                                BULK_FLUSH_BACKOFF_MAX_RETRIES_OPTION.key(),
-                                config.getBulkFlushBackoffRetries().get()));
-        if (config.getUsername().isPresent()
-                && !StringUtils.isNullOrWhitespaceOnly(config.getUsername().get())) {
-            validate(
-                    config.getPassword().isPresent()
-                            && !StringUtils.isNullOrWhitespaceOnly(config.getPassword().get()),
-                    () ->
-                            String.format(
-                                    "'%s' and '%s' must be set at the same time. Got: username '%s' and password '%s'",
-                                    USERNAME_OPTION.key(),
-                                    PASSWORD_OPTION.key(),
-                                    config.getUsername().get(),
-                                    config.getPassword().orElse("")));
-        }
-    }
-
-    private static void validate(boolean condition, Supplier<String> message) {
-        if (!condition) {
-            throw new ValidationException(message.get());
-        }
+                (Elasticsearch6Configuration) getConfiguration(context);
+        return config.getDocumentType();
     }
 
     @Override
-    public String factoryIdentifier() {
-        return "elasticsearch-6";
+    void validateConfiguration(ElasticsearchConfiguration config) {
+        super.validateConfiguration(config);
+        Elasticsearch6Configuration configuration = (Elasticsearch6Configuration) config;
+        validate(
+                !StringUtils.isNullOrWhitespaceOnly(configuration.getDocumentType()),
+                () -> String.format("'%s' must not be empty", DOCUMENT_TYPE_OPTION.key()));
     }
 
     @Override
     public Set<ConfigOption<?>> requiredOptions() {
+        Set<ConfigOption<?>> requiredOptions = super.requiredOptions();
+        requiredOptions.add(DOCUMENT_TYPE_OPTION);
         return requiredOptions;
-    }
-
-    @Override
-    public Set<ConfigOption<?>> optionalOptions() {
-        return optionalOptions;
     }
 }
